backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
pydantic
motor